Refactor + Retries for get token operation (#265)
* First refactor

* Add retries to get token operation

* Return of the IsTokenRefreshError!

* Fix CI

* Add changelog (its level of detail is subject to discussion)
