+++ autorest/retriablerequest_1.7.go
@@ -48,7 +48,8 @@ func (rr *RetriableRequest) Prepare() (err error) {
 	return err
 }
 
-func removeRequestBody(req *http.Request) {
+// RemoveRequestBody removes the request's body and sets its content length to 0
+func RemoveRequestBody(req *http.Request) {
 	req.Body = nil
 	req.ContentLength = 0
 }
