+++ autorest/retriablerequest_1.8.go
@@ -59,7 +59,8 @@ func (rr *RetriableRequest) Prepare() (err error) {
 	return err
 }
 
-func removeRequestBody(req *http.Request) {
+// RemoveRequestBody removes the request's body and sets its content length to 0
+func RemoveRequestBody(req *http.Request) {
 	req.Body = nil
 	req.GetBody = nil
 	req.ContentLength = 0
