+++ autorest/sender.go
@@ -225,7 +225,7 @@ func DoRetryForStatusCodes(attempts int, backoff time.Duration, codes ...int) Se
 				resp, err = s.Do(rr.Request())
 				// we want to retry if err is not nil (e.g. transient network failure).  note that for failed authentication
 				// resp and err will both have a value, so in this case we don't want to retry as it will never succeed.
-				if err == nil && !ResponseHasStatusCode(resp, codes...) || IsTokenRefreshError(err) {
+				if err == nil && !ResponseHasStatusCode(resp, codes...) || IsAuthenticationError(err) {
 					return resp, err
 				}
 				delayed := DelayWithRetryAfter(resp, r.Context().Done())
